Fix circular dependency
Fixes a circular dependency between OfferCodecs and LightningMessageCodecs by putting feature codecs in CommonCodecs instead.
